fixed error with long bytecodes, using long int on pow

+++ asm.h
@@ -64,7 +64,7 @@ int						valid_instruction(char *line);
 t_routine				*get_routines(char **input);
 int						get_param_size(char *param, char *cmd);
 int						get_dir_size(char *cmd);
-char					*int_to_bytecode(int val, int byte);
+char					*int_to_bytecode(long int val, int byte);
 int						get_bytecodes_count(char **split_line);
 char					**generate_output(char **input, int output_size, int prog_start);
 char					*get_parameters_bytecode(char **params, char *cmd, int line_pos, t_routine *routines);
